37: Add AM entry to /etc/hosts
Issue: https://github.com/OpenBankingToolkit/openbanking-toolkit/issues/37
Fixes and issue when building AM image for the first time. I was seeing
an issue when the Docker image was being built and amster was trying to
call AM to check to see if it had started, but of course couldn't reach
it. This fixes that issue.