830: Change spring-security auth to collect from certs
- Principal used to be collected from the session derived from the
  OAuth2 flow with the directory.
- My thinking here was to do the onboarding as the organisation and not
  the user, and to take the organisation from the certificate
(organizationIdentifier). However, this means significant changes to the
register UI as we use the session to get all of the registered apps for
the logged in user, not for that user's organisation.

Hummm. Not sure if I've taken this in the right direction or not.
Issue: https://github.com/ForgeCloud/ob-deploy/issues/830